Moved RequiresDeclaration parsing to its own file.
