+++ swift/Sources/ModuleMap/Parser+RequiresDeclarationExts.swift
